Intersected emotion data set for both Emo and Jor
Take the intersection emotion from both side and keep the row for both datasets that have the shared emotion labels from the intersection set. Note we still need to conduct train, test, valid, split on the GoEmotion dataset but not the journal. The journal is just an out of distribution test.